requirements.txt

dj-database-url==0.5.0
Django==2.0
flake8==3.5.0
gunicorn==19.9.0
whitenoise==3.3.1
psycopg2==2.7.1
django-import-export==1.0.1